🚚 move ReactorOptions file into models
