+++ src/internal/reactorList.ts
@@ -2,7 +2,7 @@ import { TextBasedChannelFields, EmojiResolvable, User } from "discord.js";
 import { makeListMessage, MessageListOptions } from "./makeListMessage";
 import { Reactor, OnEndCallback, OnReactionChangedParams, ReactorInternalOptions } from "../models/Reactor";
 import { Predicate } from "../models/Predicate";
-import { ReactorOptions } from "../models/options/ReactorOptions";
+import { ReactorOptions } from "../models/ReactorOptions";
 
 /** @internal */
 export interface ListButton<R> {
